Add cast to avoid 0 as pointer warnings

+++ test/std_ext/valmatrix.cpp
@@ -114,7 +114,7 @@ TEST(ValmatrixConstructTest, AllConstructors)
   EXPECT_FALSE(zero.size());
   const xmaho::std_ext::valmatrix<int> one {1, 1}; // mim size
   EXPECT_EQ(1, one.size());
-  EXPECT_EQ(0, one[0]);
+  EXPECT_EQ(static_cast<int>(0), one[0]);
 
   constexpr Size diff_size {2, 5};
   const xmaho::std_ext::valmatrix<int> diffsize_mat {diff_size.first, diff_size.second}; // different size
@@ -142,8 +142,8 @@ TEST(ValmatrixConstructTest, AllConstructors)
 TEST(ValmatrixSizeTest, RowAndColSizeCheck)
 {
   const xmaho::std_ext::valmatrix<int> zero {100, 0};
-  EXPECT_EQ(0, zero.row_size());
-  EXPECT_EQ(0, zero.col_size());
+  EXPECT_EQ(static_cast<std::size_t>(0u), zero.row_size());
+  EXPECT_EQ(static_cast<std::size_t>(0u), zero.col_size());
 
   const xmaho::std_ext::valmatrix<int> normal {5, 9};
   EXPECT_EQ(5, normal.row_size());
